@@ -3,7 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve --project=demo --port=4201",
+    "start": "ng serve --project=demo",
     "build:modal": "ng build ngx-bottom-sheet-modal",
     "watch:modal": "ng build ngx-bottom-sheet-modal --watch --configuration development",
     "build:phone": "ng build ngx-phonenumbers",
@@ -23,7 +23,7 @@
     "@angular/platform-server": "^19.0.0",
     "@angular/router": "^19.0.0",
     "@angular/ssr": "^19.0.1",
-    "awesome-phonenumber": "^6.12.0",
+    "awesome-phonenumber": "^7.2.0",
     "express": "^4.18.2",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
@@ -46,6 +46,6 @@
     "ng-packagr": "^19.0.0",
     "postcss": "^8.4.49",
     "tailwindcss": "^3.4.15",
-    "typescript": "~5.6.2"
+    "typescript": "^5.7.3"
   }
 }
